requirements.txt
streamlit==1.7.0
streamlit_option_menu==0.3.2
numpy==1.22.2
pandas==1.4.1
plotly==5.8.0
numpy==1.22.2
matplotlib==3.5.1
Pillow==9.0.1
protobuf==3.20.1
click==8